Created Hello_World Skill.
This skill is used as the very first step to building an Alexa skill.  All this skill will do is say hello to the user when it is spoken to.  Some of the statements are made differently when compared to other skills.  The tutorial is a bit older than most.